+++ src/layouts/CustomerLayout.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CustomerLayout = () => {
+  return <div />;
+};
+
+export default CustomerLayout;
