+++ src/layouts/ProviderLayout.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProviderLayout = () => {
+  return <div />;
+};
+
+export default ProviderLayout;
